+++ src/bonus/includes/struct_bonus.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   struct_bonus.h                                     :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sanghwal <sanghwal@student.42seoul.kr>     +#+  +:+       +#+        */
+/*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/27 11:08:43 by jgo               #+#    #+#             */
-/*   Updated: 2023/06/30 19:11:04 by sanghwal         ###   ########seoul.kr  */
+/*   Updated: 2023/07/03 14:30:07 by jgo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -246,6 +246,7 @@ struct						s_meta
 	t_thd_pool				thd_pool;
 	t_list					*spot_lights;
 	t_obj					*objs;
+	int						fd;
 };
 
 #endif
