+++ src/bonus/utils_bonus/error_bonus.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   error_bonus.c                                      :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: sanghwal <sanghwal@student.42seoul.kr>     +#+  +:+       +#+        */
+/*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/06/09 19:52:57 by jgo               #+#    #+#             */
-/*   Updated: 2023/06/30 19:35:15 by sanghwal         ###   ########seoul.kr  */
+/*   Updated: 2023/07/03 14:57:35 by jgo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -27,6 +27,10 @@ static inline void	free_all(void)
 	meta = singleton();
 	ft_lstclear(&meta->spot_lights, free);
 	objs_clear(&meta->objs, free);
+	mlx_delete_image(meta->mlx_assets.mlx, meta->mlx_assets.img);
+	mlx_terminate(meta->mlx_assets.mlx);
+	if (meta->fd != -1)
+		close(meta->fd);
 	free(meta);
 }
 
@@ -36,7 +40,8 @@ static inline void	error_parser(t_error_type type)
 	[EX_ERR] = ERR_INVALID_EX, [TYPE_ERR] = ERR_INVALID_TYPE,
 	[AMB_ERR] = ERR_AMB, [LIGHT_ERR] = ERR_LIGHT, [CAM_ERR] = ERR_CAM,
 	[SP_ERR] = ERR_SP, [PL_ERR] = ERR_PL, [CY_ERR] = ERR_CY, [CO_ERR] = ERR_CO,
-	[POINT_ERR] = ERR_POINT, [VEC_ERR] = ERR_VEC, [RGB_ERR] = ERR_RGB
+	[POINT_ERR] = ERR_POINT, [VEC_ERR] = ERR_VEC, [RGB_ERR] = ERR_RGB,
+	[THD_ERR] = ERR_THD, [OPT_ERR] = ERR_OPT
 	};
 
 	if (type >= 0 && type < sizeof(error_msgs) / sizeof(error_msgs[0])
@@ -50,12 +55,10 @@ void	error_handler(t_error_type type)
 {
 	if (type == ARGS_ERR)
 		print_error_msg(ERR_ARGS);
-	if (type == OPEN_ERR)
+	else if (type == EX_ERR)
+		print_error_msg(ERR_INVALID_EX);
+	else if (type == OPEN_ERR)
 		print_error_msg(ERR_OPEN);
-	if (type == THD_ERR)
-		print_error_msg(ERR_THD);
-	if (type == OPT_ERR)
-		print_error_msg(ERR_OPT);
 	else
 		error_parser(type);
 	exit(EXIT_FAILURE);
