+++ src/mandatory/includes/utils.h
@@ -6,7 +6,7 @@
 /*   By: jgo <jgo@student.42seoul.fr>               +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/05/27 11:09:46 by jgo               #+#    #+#             */
-/*   Updated: 2023/06/29 10:05:51 by jgo              ###   ########.fr       */
+/*   Updated: 2023/07/03 14:53:01 by jgo              ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -26,7 +26,7 @@ t_rgb		rgba_min(t_rgb a, t_rgb b);
 double		atod(char *str);
 
 // error.c
-bool		error_handler(t_error_type type);
+void		error_handler(t_error_type type);
 
 // objs.c
 void		objsadd_back(t_obj **objs, t_obj *new);
